requirements.txt
APScheduler==3.11.1
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-tailwind==1.0.3
curl_cffi==0.13.0
Django==5.2.8
django-apscheduler==0.7.0
django-crispy-forms==2.5
django-tailwind==4.4.1
frozendict==2.4.7
idna==3.11
multitasking==0.0.12
numpy==2.3.5
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pycoingecko==3.2.0
pycparser==2.23
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66
